requirements.txt
langchain==0.3.20
langchain_community==0.3.19
langchain_core==0.3.45
langchain-openai==0.3.8
faiss-cpu==1.8.0
pandas
pyprojroot
python-dotenv
PyYAML
SQLAlchemy==2.0.39
openai==1.66.3
pyhere
pypdf
openpyxl
streamlit==1.32.2